email_server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
greenlet==3.1.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
python-multipart==0.0.9
itsdangerous==2.2.0
pyyaml==6.0.1
requests==2.31.0